Extract a couple more classes
