+++ Hudl.Mjolnir/Command/AsyncCommand.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Hudl.Mjolnir.Command
+{
+    public abstract class AsyncCommand<TResult> : BaseCommand
+    {
+        public AsyncCommand(string group, string isolationKey, TimeSpan defaultTimeout) : base(group, isolationKey, defaultTimeout)
+        { }
+
+        protected internal abstract Task<TResult> ExecuteAsync(CancellationToken cancellationToken);
+    }
+}
